Utilizando Middlewares no Projeto

+++ routes/web.php
@@ -1,20 +1,23 @@
 <?php
 
-use App\Http\Controllers\ProfileController;
-use Illuminate\Support\Facades\Route;
 use \App\Http\Controllers\Admin\UserController;
+use \App\Http\Middleware\CheckIfIsAdmin;
+use \App\Http\Controllers\ProfileController;
+use Illuminate\Support\Facades\Route;
 
-Route::delete('/user/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/user/{user}', [UserController::class, 'show'])->name(('users.show'));
-Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
-Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::post('/users', [UserController::class, 'store'])->name('users.store');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
+Route::middleware('auth')->group(function (){
+    Route::delete('/user/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy')->middleware(CheckIfIsAdmin::class);
+    Route::get('/user/{user}', [UserController::class, 'show'])->name(('users.show'));
+    Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
+    Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
+    Route::post('/users', [UserController::class, 'store'])->name('users.store');
+    Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
+    Route::get('/users', [UserController::class, 'index'])->name('users.index');
+});
 
 Route::get('/', function (){
     return view('welcome');
-});
+})->name('home');
 
 Route::get('/dashboard', function () {
     return view('dashboard');
